Address review: dedup load_dataset shim, fix dtype allowlist and log label
- Extract the datasets>=4 trust_remote_code compat wrapper into a shared
  pipelinerl/domains/_datasets_compat.py; chartqa and math import it instead
  of each carrying a byte-identical copy.
- Drop the invalid "bf16" token from the vLLM dtype allowlist in launch.py
  (vLLM uses "bfloat16"; "bf16" is not an accepted --dtype value).
- Fix the unpacked-branch preprocess log to say "Unpacked microbatch".
